+++ src/bin/sks.ts
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-const FAST_PACKAGE_VERSION = '4.0.9';
+const FAST_PACKAGE_VERSION = '4.0.10';
 const args = process.argv.slice(2);
 
 try {